refactor: extrai a visão do aluno da aula para lesson.service
Move o cálculo do estado sequencial da aula (estadoDaAula) e o detalhe da aula com questões para lesson.service. O gabarito continua sem ir para o cliente. O quiz, ainda no controller, reusa o estadoDaAula importado. Comportamento preservado.
