+++ backend/src/controllers/TrailController.ts
@@ -63,6 +63,7 @@ import {
     trilhasDoUsuario,
     detalheDaTrilha,
 } from "../services/trail.service.ts";
+import { estadoDaAula, detalheDaAula } from "../services/lesson.service.ts";
 
 const QUIZ_MIN_ACERTOS = 4;
 
@@ -376,104 +377,10 @@ export const getTrail = async (req: Request, res: Response, next: NextFunction) 
     }
 };
 
-// Calcula o estado (done/current/locked) de uma aula específica na trilha.
-async function estadoDaAula(userId: string, aula: typeof lessons.$inferSelect): Promise<string> {
-    // Só aulas publicadas entram na sequência (uma aula "em breve" não trava as próximas).
-    const irmas = await db
-        .select({ id: lessons.id, moduleId: lessons.moduleId, position: lessons.position })
-        .from(lessons)
-        .where(and(eq(lessons.trailId, aula.trailId), eq(lessons.published, true)));
-    const mods = await db
-        .select({ id: modules.id, position: modules.position })
-        .from(modules)
-        .where(eq(modules.trailId, aula.trailId));
-    const ordemModulo = new Map(mods.map((m) => [m.id, m.position]));
-
-    const ordenadas = [...irmas].sort((a, b) => {
-        const pm = (ordemModulo.get(a.moduleId) ?? 0) - (ordemModulo.get(b.moduleId) ?? 0);
-        return pm !== 0 ? pm : a.position - b.position;
-    });
-
-    const concluidas = new Set(
-        (
-            await db
-                .select({ lessonId: lessonProgress.lessonId })
-                .from(lessonProgress)
-                .where(eq(lessonProgress.userId, userId))
-        ).map((p) => p.lessonId),
-    );
-
-    let achouCurrent = false;
-    for (const a of ordenadas) {
-        if (concluidas.has(a.id)) {
-            if (a.id === aula.id) return "done";
-        } else if (!achouCurrent) {
-            achouCurrent = true;
-            if (a.id === aula.id) return "current";
-        } else if (a.id === aula.id) {
-            return "locked";
-        }
-    }
-    return "locked";
-}
-
 // Retorna a aula com conteúdo e questões SEM revelar a alternativa correta.
 export const getLesson = async (req: Request, res: Response, next: NextFunction) => {
     try {
-        const userId = req.userId!;
-        const lessonId = String(req.params.id);
-
-        const [aula] = await db.select().from(lessons).where(eq(lessons.id, lessonId));
-        if (!aula) {
-            return res.status(404).json({ erro: "Aula não encontrada" });
-        }
-
-        // Aula não publicada é invisível para o aluno; admin pode visualizar (preview).
-        const admin = await ehAdmin(userId);
-        if (!aula.published && !admin) {
-            return res.status(404).json({ erro: "Aula não encontrada" });
-        }
-
-        // Admin não é travado pelo bloqueio sequencial (pode pré-visualizar qualquer aula).
-        const estado = await estadoDaAula(userId, aula);
-        if (estado === "locked" && !admin) {
-            return res.status(403).json({ erro: "Aula bloqueada. Conclua a aula anterior." });
-        }
-
-        const qs = await db
-            .select()
-            .from(questions)
-            .where(eq(questions.lessonId, lessonId))
-            .orderBy(asc(questions.position));
-        const qIds = qs.map((q) => q.id);
-        const opts = qIds.length
-            ? await db
-                  .select()
-                  .from(questionOptions)
-                  .where(inArray(questionOptions.questionId, qIds))
-                  .orderBy(asc(questionOptions.position))
-            : [];
-
-        const questoes = qs.map((q) => ({
-            id: q.id,
-            statement: q.statement,
-            position: q.position,
-            // isCorrect deliberadamente omitido: o gabarito nunca vai para o cliente.
-            options: opts
-                .filter((o) => o.questionId === q.id)
-                .map((o) => ({ id: o.id, text: o.text, position: o.position })),
-        }));
-
-        res.json({
-            id: aula.id,
-            trailId: aula.trailId,
-            moduleId: aula.moduleId,
-            title: aula.title,
-            content: aula.content,
-            contentBlocks: aula.contentBlocks ?? null,
-            state: estado,
-            questions: questoes,
-        });
+        res.json(await detalheDaAula(String(req.params.id), req.userId!));
     } catch (err) {
         next(err);
     }
